Move useAnimate to its own file
